backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
CacheControl==0.14.4
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==46.0.7
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
docutils==0.21.2
executing==2.2.0
fastapi==0.136.0
filelock==3.20.3
firebase_admin==7.4.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.55.0
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.27.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
keyring==25.5.0
kiwisolver==1.4.7
lazuardy_anfis==0.0.6
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.5
nh3==0.2.18
nltk==3.9.2
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.0.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pygad==3.4.0
Pygments==2.18.0
PyJWT==2.12.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==26.4.0
readme_renderer==44.0
regex==2025.11.3
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
safetensors==0.7.0
scikit-fuzzy==0.5.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.5.0
six==1.16.0
soupsieve==2.8
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.5
twine==5.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.44.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.17.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.21.0
